// src/types/Shikimori/Responses/Enums/ExternalLink.enum.ts
export enum ExternalLinkEnum {
  OfficialSite = 'official_site',
  Wikipedia = 'wikipedia',
  AnimeNewsNetwork = 'anime_news_network',
  MyAnimeList = 'myanimelist',
  AniDB = 'anime_db',
  WorldArt = 'world_art',
  KinoPoisk = 'kinopoisk',
  KinoPoiskHD = 'kinopoisk_hd',
  KageProject = 'kage_project',
  SubsComRu = 'subs_com_ru',
  Twitter = 'twitter',
  Anime365 = 'smotret_anime',
  ReadManga = 'readmanga',
  MangaLib = 'mangalib',
  ReManga = 'remanga',
  BakaUpdates = 'mangaupdates',
  MangaDex = 'mangadex',
  MangaFox = 'mangafox',
  MangaChan = 'mangachan',
  MangaHub = 'mangahub',
  NovelTL = 'novel_tl',
  RuRanobe = 'ruranobe',
  RanobeLib = 'ranobelib',
  NovelUpdates = 'novelupdates'
}